tools/computer_science/report_generator/requirements.txt
python-docx
markdown
beautifulsoup4
latex2mathml
regex
ahk